Implement static dispatch using our own macro
Previously, we relied on the enum_dispatch crate to implement the
InformationElement trait for the Ie enum. This was done to avoid having
to use dynamic dispatch and trait objects when dealing with a collection
of information elements. However, enum_dispatch isn't able to handle
traits with associated constants, which forced us to leave constants
that are always associated with information elements (e.g. name and ID)
out of the InformationElement trait. This commit uses our own
match_inner_ie macro that allows us to avoid trait objects while also
using associated constants.